[357] Add a confirmation before deleting a database
Before deleting a database, the action is confirmed. This confirmation
is not used for cases which result in the data being deleted due to a
pin failure.

This improves the behavior of #357 "Databases can be deleted without
password" which states "It's too easy to delete a database". Although,
this change does not prompt for a password like the request asks 
because there are other ways that the database can be removed without
a password (e.g. delete the app, enter the wrong PIN too many times, 
etc.). 

This was tested using the simulator with an iPhone XR with iOS 12.1.