[633] Fix a fresh install to not run an upgrade procedure
Version 1.6 was released Jun 14, 2015, this was the first version to
set the version number. Version 1.6 was also the first version to have
upgrade instructions. So prior versions didn't set the version number
stored in the configuration. The way the upgrade process was
implemented, fresh installs would be forced to run an upgrade.

This was fixed so that fresh installs do not run the upgrade.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1. The configuration settings were manually modified to mimic
an installation of version 1.5.2 (e.g. don't set the version but set
the pinEnabled property the old way). This triggered the upgrade from
version 1.5.2. A fresh install was tested to not perform any upgrades.