[495] Add fine grained control remembering database passwords
Fine grained control was added for remembering the database passwords.
This allows some database passwords to be remembered without
remembering all database passwords.

The on/off switch for remembering passwords was changed to a choice of
never, when configured, and always.

A new remember password switch is added to the new database and
password screens. This switch is only displayed when the "when
configured" option for remembering the password is selected.

All remembered database passwords are cleared if the never option is
selected.

A new option was added to the database swipe on the file screen for
forgetting individual remembered database passwords.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1.