[498] Fix PIN bypass when deleting or downgrading the app
In version 1.7.2 and earlier, the PIN could be bypassed by deleting
the app and then reinstalling it. When the app is deleted, this
deletes the files and configuration, but it does not necessarily
delete the keychain data for the app. After reinstalling the app, it
would trigger the upgrade path (since "[633] Fix a fresh install to
not run an upgrade procedure" was not applied). The upgrade path would
clear the PIN and then store the cleared PIN in the keychain. This
would overwrite the previous keychain PIN. However any remembered
passwords that were stored in the keychain would not be cleared. The
same data base could be reloaded and the previously remembered PIN
would be used. As a result, this allowed the PIN to be bypassed.

This was unintentionally partly fixed by "[633] Fix a fresh install to
not run an upgrade procedure" which caused the upgrade to only run if
the version actually changed. But the app was still vulnerable if the
app was maliciously downgraded to an unfixed version.

In order to protect against maliciously downgrading the version, the
app uses a different keychain service key for storing the database
passwords and keyfiles. That way old versions are unable to read the
remembered database passwords and keyfiles.

Additionally version information has been added to the keychain. This
is used to prevent against a downgrade which clears the PIN followed
by an upgrade to a patched version. The new keychain service names for
remembered database passwords and keyfiles are deleted when performing
the upgrade and also if a version rollback is detected.

In order for this fix to work, the version must be incremented on
release to a value greater than 1.7.2. Otherwise this will not protect
against a downgrade to version 1.7.2.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1. The app version was set to 1.7.3 for the manual testing. It
was tested against rolling back the version to the released version
1.7.2 and ensuring the remembered password was not usable. It was then
upgraded to the patched version which cleared the remembered password
because a downgrade was detected. The app was also deleted,
reinstalled, and reloaded with the same database that had a remembered
password. The remembered database password was checked to make sure it
was forgotten. Finally the version was updated to 1.7.4 and then it
was downgraded to 1.7.3 to make sure the keychain version actually
clears the keychain if there are downgrades in the future.