[637] Fix a vulnerability that allowed touch ID to be enabled
Previously, there was a vulnerability that allowed a malicious user to
enable touch ID and and use that to bypass the PIN. This vulnerability
required editing a file stored on the device which contained the
property for whether or not touch ID was enabled to disabled.

The touch ID property has been moved to the keychain pin service which
protects against a malicious user editing the setting.

This was manually tested using the simulator with an iPhone 6s with
iOS 12.1. When changing the touch ID setting, the property file was
verified to have the same contents.