[574] Add face ID support
Face ID support has been added and the app will prompt with a
description the first time face ID is used to authenticate.

The app detects if touch ID or face ID are present. The appropriate
text is displayed on the settings screen. If no biometrics ID is
present, then it a generic biometric ID switch is displayed.

The face ID switch has been moved into the section with PIN
protection.

The face ID text was also fixed to be properly greyed out when the PIN
is disabled.

Support for iOS prior to 11.0 cannot determine the difference between
the touch ID not being present and there not being any enrolled
fingerprints, so the generic biometric ID switch is displayed in these
cases.

This was manually tested using the simulator with an iPhone 5 with iOS
10.0 to validate the generic biometric ID switch is displayed on old
devices that do not have any biometric ID. It was tested on the
simulator with an iPhone 6s with iOS 10.0 and iOS 12.1 to make sure
the touch ID switch is displayed. It was also tested on the simulator
with an iPhone XR with iOS 12.1 to make sure the face ID switch is
displayed.