[640] Fix erase data and attempts settings to properly be disabled
The erase data on failure label was renamed and it was fixed to be 
properly greyed out when the PIN is disabled. The attempts before
erasing data was renamed and it was fixed to be properly disabled 
when the PIN is disabled or when the erase data on failure is
disabled. Both settings were moved into the PIN protection block.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1.