[648] Fix the integrated browser to close and clear cookies
The integrated web browser was fixed so that it properly closes any
open pages if the wrong PIN is entered too many times and all data is
erased. Additionally cookies and other cached files are erased.
Previously the integrated web browser was neither closed nor were the
cookies or temporary files erased.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1.