[650] Fix an incorrect attempts remaining count on the PIN screen
The attempts remaining count on the PIN screen was fixed so that it
correctly reports the number of remaining attempts.

Previously, once an incorrect PIN was entered that attempt count would
be reported even after the correct PIN was entered. 

Also previously, when the app was fully closed and reopened it would
always appear that the PIN had never been entered incorrectly even if
it had been entered incorrectly.

Both of these issues were display issues, they did not effect the 
actual counts used to erase the data.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1.