[652] Add option to cancel the set PIN screen without setting a PIN
A back button has been added to the top of the set PIN screen. This 
prevents an infinite loop of set PIN screens until the user properly
enters the same PIN twice. It allows a user to cancel the action
without actually setting a PIN followed by removing it.

This was manually tested using the simulator with an iPhone XR with
iOS 12.1.