[654] Fix a race condition that could corrupt the database when saving
The atomic flag was added when saving KeePass v2 database files. This
will use a temporary file and atomically replace the database file.

The KeePass v1 database save operation was updated to use the
DataOutputStream, which is the same as the v2 save implementation.
Temporary files are no longer created. Additionally the database file
is no longer written out followed by a separate operation which writes
the hash in the database file. Finally, the database file is created
with the data protection encryption rather than turning data protection
encryption on after the file was originally created without it.

The FileOutputStream was removed because the KeePass v1 database save
operation now uses the DataOutputStream.

This was tested manually using the simulator with an iPhone XR with
iOS 12.1. Additionally, KeePass v1 and v2 databases that were created
with MiniKeePass and they were exported and validated using KeePassX.