[626] Add screens to change the database password
The ability to change the database password has been added. A "more"
option was added to the database file selection screen for a swipe
left action. The "more" option now has the rename, forget password if
applicable, and change password.

The change password screen requires the user to enter the current
password, the remembered password cannot be used when changing the
database password.

This was tested manually using the simulator with an iPhone XR with
iOS 12.1.