[329] Hide custom in memory protected fields
The value for custom in memory protected fields are now hidden. The 
fields now acts the same way as the password field. On the view
screen, there is a eye image that displays the value. On the edit
screen, there is a wrench image that generates passwords.

This was tested manually using the simulator with an iPhone XR with
iOS 12.1.